Add tests

+++ radio/common/utils.py
@@ -1,5 +1,6 @@
 import math
 import subprocess
+import urllib.request
 from datetime import datetime
 from enum import Enum
 from functools import lru_cache, partial
@@ -7,7 +8,6 @@ from pathlib import Path
 from random import choices
 from typing import Any, Dict, List, NamedTuple, Optional, Union
 from urllib.error import URLError
-from urllib.request import urlopen
 from uuid import UUID
 
 import arrow
@@ -15,6 +15,7 @@ import marshmallow
 import mutagen
 import xmltodict
 from flask import Response, jsonify
+from marshmallow import ValidationError
 from pony.orm import commit, count, db_session, max, select, sum as db_sum
 from webargs import flaskparser
 from werkzeug.exceptions import HTTPException
@@ -22,7 +23,7 @@ from werkzeug.utils import secure_filename
 
 from radio import app
 from radio.common.schemas import RequestStatus, SongData
-from radio.models import Queue, Song, User
+from radio.database import Queue, Song
 
 register_blueprint_prefixed = partial(
     app.register_blueprint, url_prefix=app.config["SERVER_API_PREFIX"]
@@ -34,14 +35,20 @@ parser = flaskparser.FlaskParser()
 # I did not want to put this here, but since it uses make_api_response
 # there can be some tricky circular dependencies
 @parser.error_handler
-def webargs_error(error, req, schema, error_status_code, error_headers):
-    resp = make_api_response(422, "Unprocessable Entity", error.messages)
+def webargs_error(
+    error: ValidationError, req, schema, error_status_code, error_headers
+):
+    code, msg = (
+        getattr(error, "status_code", 400),
+        getattr(error, "message", "Invalid Request"),
+    )
+    resp = make_api_response(code, msg, error.messages)
     raise HTTPException(description=resp.get_data(as_text=True), response=resp)
 
 
 def make_api_response(
     status_code: int,
-    error: Union[str, bool, None],
+    error: Union[str, bool, None] = None,
     description: any = None,
     content: Dict = None,
 ) -> Response:
@@ -70,11 +77,12 @@ def get_folder_size(path: Path = Path()) -> int:
     """
     Calculate the total size of all files in the given path.
     This is cached for performance concerns.
+    Only counts files with an extension.
 
     :param path: path of folder to get size of
     :return: the total size of all files in the given path
     """
-    return sum(file.stat().st_size for file in path.rglob("*.ogg"))
+    return sum(file.stat().st_size for file in path.rglob("*.*"))
 
 
 def allowed_file_extension(filename: Path) -> bool:
@@ -86,6 +94,10 @@ def allowed_file_extension(filename: Path) -> bool:
     """
     # remove the dot
     return filename.suffix[1:].lower() in app.config["ALLOWED_EXTENSIONS"]
+
+
+class EncodeError(Exception):
+    pass
 
 
 def encode_file(input_filename: Path, remove_original: bool = False) -> Path:
@@ -102,11 +114,11 @@ def encode_file(input_filename: Path, remove_original: bool = False) -> Path:
             secure_filename(input_filename.with_suffix(".ogg").name),
         )
     )
-    subprocess.call(
+    retcode = subprocess.call(
         [
-            app.config["PATH_FFMPEG_BINARY"],
+            str(app.config["PATH_FFMPEG_BINARY"]),
             "-i",
-            input_filename,
+            str(input_filename),
             "-map_metadata",
             "0",
             "-acodec",
@@ -114,9 +126,11 @@ def encode_file(input_filename: Path, remove_original: bool = False) -> Path:
             "-q:a",
             str(app.config["SONG_QUALITY_LVL"]),
             "-vn",
-            output_path,
+            str(output_path),
         ]
     )
+    if retcode != 0:
+        raise EncodeError()
     if remove_original:
         input_filename.unlink(missing_ok=True)
     return output_path
@@ -148,7 +162,11 @@ def get_metadata(filename: Path) -> Optional[Dict[str, Any]]:
     :param Path filename: file to read tags from
     :return: dict containing music file tags
     """
-    metadata = mutagen.File(filename, easy=True)
+    try:
+        metadata = mutagen.File(filename, easy=True)
+    except:
+        app.logger.exception(f"Error loading {filename}")
+        metadata = None
     # remove files missing metadata
     if not metadata or "title" not in metadata or "artist" not in metadata:
         app.logger.warning(f"Removing {filename} due to missing metadata")
@@ -164,6 +182,7 @@ def get_metadata(filename: Path) -> Optional[Dict[str, Any]]:
     }
 
 
+@db_session
 def get_song_or_abort(song_id: Optional[UUID]) -> Song:
     """
     Returns the requested song or raises HTTPException
@@ -185,13 +204,13 @@ def insert_song(filepath: Path) -> Optional[Song]:
 
     :param Path filepath: music file to add
     """
-    print("inserting song", filepath)
+    app.logger.info(f"Inserting song: {filepath}")
     meta = get_metadata(filepath)
     if meta:
         # check dupe
         song = Song.get(artist=meta["artist"], title=meta["title"])
         if song:
-            print(filepath, "is a dupe, removing...")
+            app.logger.warning(f"{filepath} is a dupe, removing...")
             filepath.unlink(missing_ok=True)
         else:
             song = Song(
@@ -202,6 +221,9 @@ def insert_song(filepath: Path) -> Optional[Song]:
             )
             commit()
         return song
+    else:
+        app.logger.warning(f"{filepath} has no metadata, removing...")
+        filepath.unlink(missing_ok=True)
     return None
 
 
@@ -236,46 +258,31 @@ def generate_queue() -> None:
 
 
 @db_session
-def reload_songs() -> None:
+def reload_songs(path=app.config["PATH_MUSIC"]) -> None:
     """
     Keeps music directory and database in sync.
     This is done by removing songs that exist in the database, but not on the filesystem.
     Also adds any songs found in the filesystem that are not present in the database.
     """
-    os_songs = [f.name for f in app.config["PATH_MUSIC"].glob("*.ogg")]
+    songs_filesystem = [f.name for f in path.glob("*.ogg")]
     # insert all files if no songs in database
     if count(s for s in Song) <= 0:
-        for filename in os_songs:
-            insert_song(Path(app.config["PATH_MUSIC"], filename))
+        for filename in songs_filesystem:
+            insert_song(path / filename)
     # generate queue if it is empty
     if count(s for s in Queue) <= 0:
         generate_queue()
 
-    db_songs = select(song.filename for song in Song)[:]
+    songs_db = select(song.filename for song in Song)[:]
 
-    songs_to_add = []
-    songs_to_remove = []
-
-    for song in os_songs:
-        if song not in db_songs:
-            songs_to_add.append(song)
-
-    for song in db_songs:
-        if song not in os_songs:
-            songs_to_remove.append(song)
+    songs_to_add = [song for song in songs_filesystem if song not in songs_db]
+    songs_to_remove = [song for song in songs_db if song not in songs_filesystem]
 
     for song in songs_to_remove:
-        db_song = Song.get(filename=song)
-        if db_song:
-            queue_song = Queue.get(song=db_song)
-            if queue_song:
-                queue_song.delete()
-            for user in User.select():
-                user.favourites.remove(db_song)
-            db_song.delete()
+        Song.get(filename=song).delete()
 
-    print("songs to add", songs_to_add)
-    print("songs to remove", songs_to_remove)
+    app.logger.info(f"Adding songs: {songs_to_add}")
+    app.logger.info(f"Removing songs: {songs_to_remove}")
 
     for filename in songs_to_add:
         insert_song(Path(app.config["PATH_MUSIC"], filename))
@@ -289,9 +296,8 @@ def next_song() -> Optional[Path]:
     :return: path to song
     """
     generate_queue()
-    try:
-        queue_entry = Queue.select().sort_by(Queue.id)[:1][0]
-    except IndexError:
+    queue_entry = Queue.select().sort_by(Queue.id).first()
+    if not queue_entry:
         return None
     song = Song[queue_entry.song.id]
     song.playcount += 1
@@ -324,9 +330,8 @@ def queue_status(song: Song) -> QueueStatus:
     :param song: song to get status of
     :return: queue details for given song
     """
-    song_queue = Queue.select(lambda s: s.song.id == song.id)[:]
+    song_queue = Queue.select(lambda s: s.song.id == song.id).sort_by(Queue.id).first()
     if song_queue:
-        song_queue = song_queue[-1]
         req_type = QueueType.USER if song_queue.requested else QueueType.NORMAL
         return QueueStatus(queued=True, type=req_type, time=song_queue.added)
     return QueueStatus(queued=False, type=QueueType.NONE, time=None)
@@ -395,7 +400,7 @@ def parse_status(url: str) -> dict:
     """
     result = {"Online": False, "Current Song": ""}  # Assume False by default
     try:
-        xml = urlopen(url).read()
+        xml = urllib.request.urlopen(url).read()
     except URLError:
         return result
     try:
@@ -423,16 +428,14 @@ def parse_status(url: str) -> dict:
                 result[tmp[0]] = tmp[1].strip()
         result["Online"] = True
         result["Current Song"] = xml_dict.get("title", "") or ""
-    except UnicodeDecodeError:
-        # we have runes, but we know we are online. This should not even be
-        # possible (requests.get.content)
-        result["Online"] = True
-        # Erase the bad stuff. However, keep in mind stream title can do this (anything user input...)
+    except:
+        # fallback for any unexpected errors
+        result["Online"] = False
         result["Current Song"] = ""
     return result
 
 
-def filter_default_webargs(args: dict, **kwargs: dict) -> dict:
+def filter_default_webargs(args: marshmallow.Schema, **kwargs: any) -> dict:
     """
     Returns a dict containing only arguments that have non-default values set
 
@@ -440,17 +443,15 @@ def filter_default_webargs(args: dict, **kwargs: dict) -> dict:
     :param kwargs: The arguments to filter
     :return: dict containing only arguments that have non-default values set
     """
-    if isinstance(args, marshmallow.Schema):
-        args = args.fields
     res = {}
     for kwarg, val in kwargs.items():
-        if kwarg in args:
-            webarg = args[kwarg]
-            if webarg.default is marshmallow.missing:
-                webarg_def = webarg.missing
+        if kwarg in args.fields:
+            field = args.fields[kwarg]
+            if field.default is marshmallow.missing:
+                default = field.missing
             else:
-                webarg_def = webarg.default
-            if val != webarg_def:
+                default = field.default
+            if val != default:
                 res[kwarg] = val
     return res
 
@@ -458,13 +459,8 @@ def filter_default_webargs(args: dict, **kwargs: dict) -> dict:
 def get_nonexistant_path(fname: Path):
     """
     Get the path to a filename which does not exist by incrementing path.
-
-    Examples
-    --------
-    >>> get_nonexistant_path('/etc/issue')
-    '/etc/issue-1'
-    >>> get_nonexistant_path('whatever/doesnt-exist-yet.py')
-    'whatever/doesnt-exist-yet.py'
+    :param fname: A path to get a free filename for
+    :return: Path to a nonexistant file
     """
     if not fname.exists():
         return fname
